+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - Allow sprite frames to have different sizes
 - Fix missing color in index-alpha sprite extract
 - Add --no-alpha flag to sprite extract
+- Add --no-alpha flag to wad extract and make '{'-prefixed textures transparent by default
 
 # v1.5.0
 - Add "map neat" entity preprocessor command